src/mcpserver/flightbooking/pyproject.toml
[project]
name = "flightbooking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiodns>=3.5.0",
    "aiohttp>=3.12.15",
    "mcp[cli]>=1.12.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
